abstracts list of all fish and angler search to their own components
