+++ src/containers/index.js
@@ -1,20 +1,11 @@
 import React, {Component, Fragment} from 'react';
 import {connect} from 'react-redux';
-import {DebounceInput} from 'react-debounce-input';
 import {fetchAllFish} from '../redux/actions';
+import AnglerSearch from '../containers/anglerSearch';
 import Column from '../components/column';
-import FishList from '../components/fishList';
+import AllFishList from '../components/allFish';
 
 class Index extends Component {
-  constructor() {
-    super();
-
-    this.state = {
-      anglerSearch: '',
-      currentAngler: []
-    }
-  }
-
   componentWillMount() {
     this.props.fetchAllFish();
   }
@@ -22,44 +13,8 @@ class Index extends Component {
   render() {
     return (
       <div className='row'>
-        <Column className='col-md-4'>
-          {/*LISTS ALL FISH*/}
-          <h3>All Catfish</h3>
-          <FishList catfish={this.props.catfish.all} />
-        </Column>
-        <Column className='col-md-4'>
-          {/*LISTS TOP FIVE FISH BY ANGLER*/}
-          <h3>Catfish by Angler</h3>
-          <DebounceInput
-            minLength={2}
-            debounceTimeout={500}
-            value={this.state.anglerSearch}
-            onChange={event => {
-              const currentAngler = this.props.catfish.all.filter(fish =>
-                fish.angler.toLowerCase() === event.target.value.toLowerCase()
-              ).map(fish => {
-                return {
-                  ...fish,
-                  weight: fish.length * fish.girth * fish.girth / 800
-                }
-              }).sort((a, b) => {
-                return b.weight - a.weight;
-              });
-              this.setState({
-                anglerSearch: event.target.value,
-                currentAngler
-              })
-            }}
-          />
-            {this.state.currentAngler.length < 1 ? (
-              <p>Type an angler's name above</p>
-            ) : (
-              <Fragment>
-                <h3>{this.state.anglerSearch}'s Top 5 fish</h3>
-                <FishList catfish={this.state.currentAngler.slice(0, 5)} />
-              </Fragment>
-            )}
-        </Column>
+        <AllFishList catfish={this.props.catfish.all} />
+        <AnglerSearch />
         <Column className='col-md-4'>
         </Column>
       </div>
